requirements.txt
sounddevice>=0.4.6
soundfile>=0.12
numpy>=1.26
openai>=1.50
python-dotenv>=1.0
PySide6>=6.6
